add custom hooks useConstants
